+++ R/MetaAnalysis.R
@@ -25,12 +25,13 @@
 #' @param data      A data frame containing either normal, skew-normal, custom parametric, or grid
 #'                  likelihood data. One row per database.
 #' @param labels    A vector of labels for the data sources.
-#' @param estimate  The meta-analytic estimate as created using either ['computeFixedEffectMetaAnalysis()`]
-#'                  or [`computeBayesianMetaAnalysis()`] function.
+#' @param estimate  The meta-analytic estimate as created using either [computeFixedEffectMetaAnalysis]
+#'                  or [computeBayesianMetaAnalysis] function.
 #' @param xLabel    The label on the x-axis: the name of the effect estimate.
 #' @param summaryLabel The label for the meta-analytic estimate.
 #' @param limits    The limits of the effect size axis.
 #' @param alpha     The alpha (expected type I error).
+#' @param showPredictionInterval Show the prediction interval (for random effects models).
 #' @param showLikelihood Show the likelihood curve for each estimate?
 #' @param fileName  Name of the file where the plot should be saved, for example 'plot.png'. See the
 #'                  function [ggplot2::ggsave] ifor supported file formats.
@@ -50,11 +51,10 @@
 #'     modelType = "cox"
 #'   )
 #'   cyclopsFit <- Cyclops::fitCyclopsModel(cyclopsData)
-#'   approximation <- approximateLikelihood(cyclopsFit, parameter = "x", approximation = "custom")
+#'   approximation <- approximateLikelihood(cyclopsFit, parameter = "x", approximation = "grid with gradients")
 #'   return(approximation)
 #' }
 #' approximations <- lapply(populations, fitModelInDatabase)
-#' approximations <- do.call("rbind", approximations)
 #'
 #' # At study coordinating center, perform meta-analysis using per-site approximations:
 #' estimate <- computeBayesianMetaAnalysis(approximations)
@@ -70,6 +70,7 @@ plotMetaAnalysisForest <- function(data,
                                    summaryLabel = "Summary",
                                    limits = c(0.1, 10),
                                    alpha = 0.05,
+                                   showPredictionInterval = TRUE,
                                    showLikelihood = TRUE,
                                    fileName = NULL) {
   if (is.numeric(data)) {
@@ -81,10 +82,10 @@ plotMetaAnalysisForest <- function(data,
   }
 
   if (length(data) != length(labels)) {
-    abort("The number of labels should be equal to the number of approximatins in `data`")
+    abort("The number of labels should be equal to the number of approximations in `data`")
   }
 
-  d1 <- data.frame(
+  d1 <- tibble(
     logRr = -100,
     logLb95Ci = -100,
     logUb95Ci = -100,
@@ -96,7 +97,7 @@ plotMetaAnalysisForest <- function(data,
       approximation = approximation,
       alpha = alpha
     ))
-    return(data.frame(
+    return(tibble(
       logRr = ci$logRr,
       logLb95Ci = log(ci$lb),
       logUb95Ci = log(ci$ub),
@@ -104,10 +105,12 @@ plotMetaAnalysisForest <- function(data,
     ))
   }
   d2 <- lapply(data, getEstimate)
-  d2 <- do.call(rbind, d2)
-  d2$label <- labels
+  d2 <- bind_rows(d2) |>
+    mutate(label = labels)
+
   if ("rr" %in% colnames(estimate)) {
-    d3 <- data.frame(
+    # Estimate produced by computeFixedEffectMetaAnalysis
+    d3 <- tibble(
       logRr = estimate$logRr,
       logLb95Ci = log(estimate$lb),
       logUb95Ci = log(estimate$ub),
@@ -115,97 +118,64 @@ plotMetaAnalysisForest <- function(data,
       label = summaryLabel
     )
   } else if ("mu" %in% colnames(estimate)) {
-    d3 <- data.frame(
-      logRr = estimate$mu,
-      logLb95Ci = estimate$mu95Lb,
-      logUb95Ci = estimate$mu95Ub,
-      type = "ma",
-      label = sprintf("%s (tau = %.2f)", summaryLabel, estimate$tau)
+    # Estimate produced by computeBayesianMetaAnalysis
+    d3 <- tibble(
+      logRr = c(estimate$logRr, NA),
+      logLb95Ci = c(estimate$mu95Lb, NA),
+      logUb95Ci = c(estimate$mu95Ub, NA),
+      type = c("ma", "maSub"),
+      label = c("Bayesian random effects", sprintf("\u03C4 = %.2f (%.2f - %.2f)", estimate$tau, estimate$tau95Lb, estimate$tau95Ub))
     )
+    if (showPredictionInterval) {
+      d3 <- bind_rows(
+        d3,
+        tibble(
+          logRr = NA,
+          logLb95Ci = estimate$predictionInterval95Lb,
+          logUb95Ci = estimate$predictionInterval95Ub,
+          type = "pi",
+          label = "Prediction interval"
+        )
+      )
+    }
   } else {
-    d3 <- data.frame(
-      logRr = estimate$logRr,
-      logLb95Ci = log(estimate$logLb95Ci),
-      logUb95Ci = log(estimate$logUb95Ci),
-      type = "ma",
-      label = summaryLabel
-    )
+    stop("Unknown summary estimate type")
   }
-  d <- rbind(d1, d2, d3)
-  d$y <- seq(nrow(d), 1)
-  # d$name <- factor(d$name, levels = c(d3$name, rev(as.character(labels)), "Source"))
+  d <- bind_rows(d1, d2, d3)
+  d <- d |>
+    mutate(y = if_else(.data$type == "maSub", 0.5, 1)) |>
+    mutate(y = sum(.data$y) - cumsum(.data$y) + 1)
+  maEstimates <- d |>
+    filter(.data$type == "ma", !is.na(.data$logRr))
+  diamondData <- tibble(
+    x = exp(c(maEstimates$logLb95Ci, maEstimates$logRr, maEstimates$logUb95Ci, maEstimates$logRr)),
+    y = c(maEstimates$y, maEstimates$y + 0.2, maEstimates$y, maEstimates$y - 0.2),
+    group = rep(maEstimates$y, 4)
+  )
+
+  maBoundaryY <- d |>
+    filter(.data$type == "ma") |>
+    summarise(max(.data$y)) |>
+    pull() + 0.5
 
   # ggplot puts whisker for infinite values, but not large values:
   plotD <- d
   plotD$logLb95Ci[is.infinite(plotD$logLb95Ci)] <- -10
   plotD$logUb95Ci[is.infinite(plotD$logUb95Ci)] <- 10
+  plotD <- plotD |>
+    filter(.data$type != "ma")
 
-  rowHeight <- 0.8
+  rowHeight <- 0.5
   breaks <- c(0.1, 0.25, 0.5, 1, 2, 4, 6, 8, 10)
   yLimits <- c(min(d$y) - rowHeight / 2, max(d$y) + rowHeight / 2)
-  p <- ggplot2::ggplot(plotD, ggplot2::aes(x = exp(.data$logRr), y = .data$y))
-
-  if (showLikelihood) {
-    yLimit <- -qchisq(1 - alpha, df = 1) / 2
-    transformY <- function(coords, yOffset) {
-      coords$y <- coords$y - max(coords$y)
-      coords <- coords[coords$y >= yLimit, ]
-      coords$y <- rowHeight / 2 - (coords$y * rowHeight / yLimit) + yOffset
-      coords$y2 <- yOffset - rowHeight / 2
-      return(coords)
-    }
-    getCoords <- function(i) {
-      coords <- getLikelihoodCoordinates(data[[i]], limits, verbose = FALSE)
-      coords <- transformY(coords, length(data) - i + 2)
-      coords$group <- length(data) - i + 2
-      return(coords)
-    }
-    llData1 <- lapply(seq_along(data), getCoords)
-    llData1 <- do.call(rbind, llData1)
-
-    if ("rr" %in% colnames(estimate)) {
-      llData2 <- getLikelihoodCoordinates(estimate, limits, verbose = FALSE)
-      llData2 <- transformY(llData2, 1)
-      llData2$group <- 1
-    } else {
-      trace <- attr(estimate, "traces")[, 1]
-      trace <- trace[trace > log(limits[1]) & trace < log(limits[2])]
-      coords <- density(trace, from = log(limits[1]), to = log(limits[2]), n = 100)
-      llData2 <- data.frame(x = coords$x, y = log(coords$y))
-      llData2 <- transformY(llData2, 1)
-      llData2$group <- 1
-    }
-
-    llData <- rbind(llData1, llData2)
-    uniqueYs <- unique(llData$y2)
-    nSteps <- 10
-    ys <- rep(uniqueYs, nSteps) + rep((rowHeight / 2) * ((seq_len(nSteps) - 1) / nSteps), each = length(uniqueYs))
-    fadeData <- data.frame(
-      xmin = limits[1],
-      xmax = limits[2],
-      ymin = ys,
-      ymax = ys + rowHeight / 2 / nSteps,
-      alpha = 1 - rep(seq_len(nSteps) / nSteps, each = length(uniqueYs)),
-      logRr = 0,
-      y = 1
-    )
-    p <- p + ggplot2::geom_ribbon(
-      ggplot2::aes(x = exp(.data$x), ymax = .data$y, ymin = .data$y2, group = .data$group),
-      fill = "#0000AA",
-      alpha = 0.25,
-      data = llData
-    ) +
-      ggplot2::geom_rect(ggplot2::aes(xmin = .data$xmin, xmax = .data$xmax, ymin = .data$ymin, ymax = .data$ymax, alpha = .data$alpha), fill = "#FFFFFF", data = fadeData)
-  }
-  p <- p +
-    ggplot2::geom_vline(xintercept = breaks, colour = "#AAAAAA", lty = 1, size = 0.2) +
-    ggplot2::geom_vline(xintercept = 1, size = 0.5) +
-    ggplot2::geom_errorbarh(ggplot2::aes(
-      xmin = exp(.data$logLb95Ci),
-      xmax = exp(.data$logUb95Ci)
-    ), height = 0.15) +
-    ggplot2::geom_point(size = 3, shape = 23, ggplot2::aes(fill = .data$type)) +
-    ggplot2::scale_fill_manual(values = c("#000000", "#000000", "#FFFFFF")) +
+  rightPlot <- ggplot2::ggplot(plotD, ggplot2::aes(x = exp(.data$logRr), y = .data$y)) +
+    ggplot2::geom_rect(xmin = -10, xmax = 10, ymin = 0, ymax = maBoundaryY, size = 0, fill = "#69AED5", alpha = 0.25, data = tibble(logRr = 1, y = 1)) +
+    ggplot2::geom_segment(ggplot2::aes(x = .data$x, y = .data$y, xend = .data$x, yend = .data$yend), color = "#AAAAAA",  size = 0.2, data = data.frame(x = breaks, y = 0, yend = max(d$y) - 0.5)) +
+    ggplot2::geom_segment(ggplot2::aes(x = .data$x, y = .data$y, xend = .data$x, yend = .data$yend), size = 0.5, data = data.frame(x = 1, y = 0, yend = max(d$y) - 0.5)) +
+    ggplot2::geom_hline(ggplot2::aes(yintercept = max(d$y) - 0.5)) +
+    ggplot2::geom_errorbarh(ggplot2::aes(xmin = exp(.data$logLb95Ci), xmax = exp(.data$logUb95Ci)), height = 0.15) +
+    ggplot2::geom_point(size = 3, shape = 16) +
+    ggplot2::geom_polygon(ggplot2::aes(x = .data$x, y = .data$y, group = .data$group), data = diamondData) +
     ggplot2::scale_x_continuous(xLabel, trans = "log10", breaks = breaks, labels = breaks) +
     ggplot2::coord_cartesian(xlim = limits, ylim = yLimits) +
     ggplot2::theme(
@@ -214,52 +184,57 @@ plotMetaAnalysisForest <- function(data,
       panel.background = ggplot2::element_blank(),
       legend.position = "none",
       panel.border = ggplot2::element_blank(),
+      axis.text.x = ggplot2::element_text(color = "black"),
       axis.text.y = ggplot2::element_blank(),
+      axis.ticks.y = ggplot2::element_blank(),
       axis.title.y = ggplot2::element_blank(),
-      axis.ticks = ggplot2::element_blank(),
-      plot.margin = grid::unit(c(0, 0, 0.1, 0), "lines")
+      axis.line.x.bottom = ggplot2::element_line(),
+      plot.margin = ggplot2::unit(c(0, 0, 0, -0.19), "lines")
     )
-
-  # p
+  rightPlot
   d$logLb95Ci[is.infinite(d$logLb95Ci)] <- NA
   d$logUb95Ci[is.infinite(d$logUb95Ci)] <- NA
   d$logRr[exp(d$logRr) < limits[1] | exp(d$logRr) > limits[2]] <- NA
-  labels <- sprintf("%0.2f (%0.2f - %0.2f)", exp(d$logRr), exp(d$logLb95Ci), exp(d$logUb95Ci))
-  labels <- gsub("NA", "", labels)
-  labels <- gsub(" \\( - \\)", "-", labels)
-  labels <- data.frame(
-    y = rep(d$y, 2),
-    x = rep(1:2, each = nrow(d)),
-    label = c(as.character(d$label), labels),
-    stringsAsFactors = FALSE
+  estimateLabels <- sprintf("%0.2f", exp(d$logRr))
+  estimateLabels <- gsub("NA", "-", estimateLabels)
+  estimateLabels[d$type %in% c("maSub", "pi", "header")] <- ""
+  intervalLabels <- sprintf("(%0.2f - %0.2f)", exp(d$logLb95Ci), exp(d$logUb95Ci))
+  intervalLabels <- gsub("NA", "", intervalLabels)
+  intervalLabels <- gsub("\\( - \\)", "", intervalLabels)
+  intervalLabels[d$type %in% c("maSub", "header")] <- ""
+
+  textTable <- data.frame(
+    y = rep(d$y, 3),
+    x = rep(c(1, 2, 2.2), each = nrow(d)),
+    label = c(as.character(d$label), estimateLabels, intervalLabels),
+    fontface = rep(if_else(d$type %in% c("ma", "header", "pi"), "bold", "plain"), 3)
   )
-  labels$label[nrow(d) + 1] <- paste(xLabel, "(95% CI)")
-  data_table <- ggplot2::ggplot(labels, ggplot2::aes(
-    x = .data$x,
-    y = .data$y,
-    label = .data$label
-  )) +
-    ggplot2::geom_text(size = 4, hjust = 0, vjust = 0.5) +
-    ggplot2::geom_hline(ggplot2::aes(yintercept = nrow(d) - 0.5)) +
-    ggplot2::scale_y_continuous(limits = yLimits) +
+  textTable$label[nrow(d) + 1] <- paste(xLabel, "(95% CI)")
+  xLimits <- c(1, 2.75)
+  widths <- c(1.5, 1)
+  width <- 7
+  leftPlot <- ggplot2::ggplot(textTable, ggplot2::aes(x = .data$x, y = .data$y, label = .data$label)) +
+    ggplot2::geom_rect(xmin = -10, xmax = 10, ymin = 0, ymax = maBoundaryY, size = 0, fill = "#69AED5", alpha = 0.25, data = tibble(x = 1, y = 1, label = "NA")) +
+    ggplot2::geom_text(ggplot2::aes(fontface = .data$fontface), size = 4, hjust = 0, vjust = 0.5) +
+    ggplot2::geom_hline(ggplot2::aes(yintercept = max(d$y) - 0.5)) +
+    ggplot2::labs(x = "", y = "") +
+    ggplot2::coord_cartesian(xlim = xLimits, ylim = yLimits) +
     ggplot2::theme(
       panel.grid.major = ggplot2::element_blank(),
       panel.grid.minor = ggplot2::element_blank(),
+      panel.background = ggplot2::element_blank(),
       legend.position = "none",
       panel.border = ggplot2::element_blank(),
-      panel.background = ggplot2::element_blank(),
-      axis.text.x = ggplot2::element_text(colour = "white"),
+      axis.text.x = ggplot2::element_text(color = "white"),
       axis.text.y = ggplot2::element_blank(),
-      axis.ticks = ggplot2::element_line(colour = "white"),
-      plot.margin = grid::unit(c(0, 0, 0.1, 0), "lines")
-    ) +
-    ggplot2::labs(x = "", y = "") +
-    ggplot2::coord_cartesian(xlim = c(1, 3))
-
-  plot <- gridExtra::grid.arrange(data_table, p, ncol = 2)
-
+      axis.ticks = ggplot2::element_blank(),
+      axis.title.y = ggplot2::element_blank(),
+      axis.line.x.bottom = ggplot2::element_line(),
+      plot.margin = ggplot2::unit(c(0, 0, 0, 0), "lines")
+    )
+  plot <- gridExtra::grid.arrange(leftPlot, rightPlot, ncol = 2, widths = widths, padding = ggplot2::unit(0, "line"))
   if (!is.null(fileName)) {
-    ggplot2::ggsave(fileName, plot, width = 7, height = 1 + length(data) * 0.3, dpi = 400)
+    ggplot2::ggsave(fileName, plot, width = width, height = 1 + max(d$y) * 0.3, dpi = 300)
   }
   invisible(plot)
 }